PDDA-474 Moving getCourtSites() out of CourtService into AbstractService
